refactor: 도메인에 맞게 카드더미 네이밍 변경
